refactor: get rid of proposalDetail store
